backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0,<1.0.0",
    "sqlalchemy>=2.0.46,<3.0.0",
    "uvicorn>=0.40.0,<1.0.0",
    "pydantic-settings>=2.13.0,<3.0.0",
    "psycopg2-binary>=2.9.11,<3.0.0",
    "alembic>=1.18.4,<2.0.0",
    "langgraph>=0.4.0",
    "kubernetes>=32.0.0",
    "boto3>=1.35.0",
]


[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]
dev = [
    "ruff>=0.11.0",
    "mypy>=1.14.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py313"
line-length = 100
exclude = ["migration/versions/"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
exclude = ["migration/"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
disable_error_code = ["arg-type"]

[[tool.mypy.overrides]]
module = "app.workflow.*"
disallow_untyped_calls = false

[[tool.mypy.overrides]]
module = ["langgraph.*", "kubernetes.*", "pytest.*", "pytest_asyncio.*", "_pytest.*", "boto3.*", "botocore.*"]
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
